+++ amazon/s3/client/src/main/kotlin/org/http4k/connect/amazon/s3/action/PutObject.kt
@@ -6,6 +6,7 @@ import org.http4k.connect.Http4kConnectAction
 import org.http4k.connect.amazon.core.model.Tag
 import org.http4k.connect.amazon.s3.S3BucketAction
 import org.http4k.connect.amazon.s3.model.BucketKey
+import org.http4k.connect.amazon.s3.model.StorageClass
 import org.http4k.connect.asRemoteFailure
 import org.http4k.core.Headers
 import org.http4k.core.Method.PUT
@@ -20,8 +21,12 @@ data class PutObject(
     val content: InputStream,
     val headers: Headers = emptyList(),
     val tags: List<Tag> = emptyList(),
+    val storageClass: StorageClass? = null,
 ) : S3BucketAction<Unit> {
-    override fun toRequest() = Request(PUT, Uri.of("/$key")).headers(headers + headersFor(tags)).body(content)
+
+    override fun toRequest() = Request(PUT, Uri.of("/$key"))
+        .headers(headers + headersFor(tags) + headersFor(storageClass))
+        .body(content)
 
     override fun toResult(response: Response) = with(response) {
         when {
@@ -31,8 +36,14 @@ data class PutObject(
     }
 }
 
-private fun headersFor(tags: Collection<Tag>) = if (tags.isEmpty()) {
+private fun headersFor(tags: List<Tag>) = listOfNotNull(headerFor(tags))
+
+fun headerFor(tags: List<Tag>) = tags
+    .takeIf { it.isNotEmpty() }
+    .let { "x-amz-tagging" to tags.joinToString("&") { (key, value) -> "$key=$value" } }
+
+private fun headersFor(storageClass: StorageClass?) = if (storageClass == null) {
     emptyList()
 } else {
-    listOf(Pair("x-amz-tagging", tags.joinToString("&") { (key, value) -> "$key=$value" }))
+    listOf("x-amz-storage-class" to storageClass.toString())
 }