+++ amazon/s3/fake/src/main/kotlin/org/http4k/connect/amazon/s3/FakeS3.kt
@@ -10,6 +10,7 @@ import org.http4k.connect.amazon.s3.endpoints.bucketGetKey
 import org.http4k.connect.amazon.s3.endpoints.bucketHeadBucket
 import org.http4k.connect.amazon.s3.endpoints.bucketHeadKey
 import org.http4k.connect.amazon.s3.endpoints.bucketListObjectsV2
+import org.http4k.connect.amazon.s3.endpoints.bucketPostKey
 import org.http4k.connect.amazon.s3.endpoints.bucketPutBucket
 import org.http4k.connect.amazon.s3.endpoints.bucketPutKey
 import org.http4k.connect.amazon.s3.endpoints.copyKey
@@ -20,6 +21,7 @@ import org.http4k.connect.amazon.s3.endpoints.globalPutBucket
 import org.http4k.connect.amazon.s3.endpoints.pathBasedBucketDeleteKey
 import org.http4k.connect.amazon.s3.endpoints.pathBasedBucketGetKey
 import org.http4k.connect.amazon.s3.endpoints.pathBasedBucketHeadKey
+import org.http4k.connect.amazon.s3.endpoints.pathBasedBucketPostKey
 import org.http4k.connect.amazon.s3.endpoints.pathBasedBucketPutKey
 import org.http4k.connect.amazon.s3.endpoints.pathBasedCopyKey
 import org.http4k.connect.amazon.s3.endpoints.subdomain
@@ -48,6 +50,7 @@ class FakeS3(
         isS3 bind routes(
             pathBasedCopyKey(buckets, bucketContent, clock),
             pathBasedBucketGetKey(buckets, bucketContent),
+            pathBasedBucketPostKey(buckets, bucketContent),
             pathBasedBucketPutKey(buckets, bucketContent, clock),
             pathBasedBucketDeleteKey(buckets, bucketContent),
             pathBasedBucketHeadKey(buckets, bucketContent),
@@ -61,6 +64,7 @@ class FakeS3(
             bucketHeadBucket(buckets),
             bucketHeadKey(buckets, bucketContent),
             bucketGetKey(buckets, bucketContent),
+            bucketPostKey(buckets, bucketContent),
             bucketPutKey(buckets, bucketContent, clock),
             bucketDeleteKey(buckets, bucketContent),
             bucketPutBucket(buckets),