+++ amazon/s3/fake/src/main/kotlin/org/http4k/connect/amazon/s3/endpoints/headKey.kt
@@ -5,7 +5,6 @@ import org.http4k.connect.storage.Storage
 import org.http4k.core.Method.HEAD
 import org.http4k.core.Request
 import org.http4k.core.Response
-import org.http4k.core.Status.Companion.NOT_FOUND
 import org.http4k.core.Status.Companion.OK
 import org.http4k.routing.bind
 import org.http4k.routing.path
@@ -25,10 +24,8 @@ fun bucketHeadKey(
     bucket: String,
     bucketContent: Storage<BucketKeyContent>,
     req: Request
-) = (buckets[bucket]
-    ?.let {
-        Response(
-            bucketContent["${bucket}-${req.path("bucketKey")!!}"]
-                ?.let { OK } ?: NOT_FOUND)
-    }
-    ?: invalidBucketNameResponse())
+): Response {
+    if (buckets[bucket] == null) return invalidBucketNameResponse()
+    val obj = bucketContent["${bucket}-${req.path("bucketKey")!!}"] ?: return invalidBucketKeyResponse()
+    return Response(OK).headers(getHeadersWithoutXHttp4kPrefix(obj))
+}