+++ amazon/s3/fake/src/main/kotlin/org/http4k/connect/amazon/s3/model.kt
@@ -1,5 +1,6 @@
 package org.http4k.connect.amazon.s3
 
+import org.http4k.connect.amazon.core.model.Tag
 import org.http4k.connect.amazon.s3.model.BucketKey
 import org.http4k.connect.amazon.s3.model.BucketName
 import org.http4k.core.Headers
@@ -12,7 +13,8 @@ data class BucketKeyContent(
     val key: BucketKey,
     val content: String,
     val modified: Instant,
-    val headers: Headers
+    val headers: Headers,
+    val tags: List<Tag>
 ) {
     val size = content.length
 }
@@ -22,4 +24,4 @@ data class ListBucketResult(val bucketName: String, val keys: List<BucketKeyCont
     val maxKeys = Integer.MAX_VALUE
 }
 
-data class S3Error(val code: String, val resource: String? = "") : ViewModel
+data class S3Error(val code: String, val resource: String? = "", val message: String) : ViewModel