+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This list is not intended to be all-encompassing - it will document major and br
 changes with their rationale when appropriate. Given version `A.B.C.D`, breaking changes are to be expected in version number increments where changes in the `A` or `B` sections:
 
 ### v5.17.1.0 (uncut)
+- **http4k-connect-amazon-s3-*** - Support all storage classes, including the restore lifecycle for glacier. H/T @oharaandrew314
 - **http4k-connect-*** - Upgrade dependencies.
 
 ### v5.17.0.2